Algorithms in separate modules based on category
